bin/test-lxd-storage-vm: Check root disk block device size rather than filesystem size
Means we don't have to wait for it to be resized by cloud-init, and we don't have to account for the /boot partition.
Ultimately all this check is doing is to check in a storage-driver agnostic way that the disks are at the expected size.
Using `df` of the filesystem was just a proxy to the root disk's block device size, and not a reliable one.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>