bin/test-lxd-network-ovn: Use -w flag rather than -W flag with ping commands
According to the ping manpage the -w does:

 -w deadline
           Specify a timeout, in seconds, before ping exits regardless of how many packets have been
           sent or received. In this case ping does not stop after count packet are sent, it waits
           either for deadline expire or until count probes are answered or for some error
           notification from network.

This seems preferable to use in an attempt to address flakiness in the nested vlan ping
tests on Jenkins (which I do not experience when running in a local Focal VM), as it will
allow multiple packets to be sent up to a certain timeout or after specific number of replies
have been receieved (usually 1 for our tests just to check the target is reachable).

This should allow us to workaround any slowness in OVN setting up the ports internally.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>