bin/maas-kernel: Tweak the cpu set passed to isolcpus=
This tweaks the set of CPUs passed to the isolcpus= kernel option.

This setting is taylor-made for lantea which has 2 sockets, with 6 physical
cores each, and 2 hyper threads per core.

With the CPU set in this commit the kernel won't use the second socket at all,
and all those 6 physical cores can be reserved for benchmarking.

We might need to make this value dynamic in the future, and not
hard-coded/tailormade, but this is good enough for now.

Signed-off-by: Free Ekanayaka <free@ekanayaka.io>