Fixed hardcoded root partition device name sda2
The root partition may not be /dev/sda2, for example if set the boot disk as the
second disk, the root partition device name will be /dev/sdb2. It may be
/dev/vda2 if not using san disk. The fix figured out the root partition device
UUID and set the UUID as the root, the root partition device UUID keeps no
change.

Depends on lxc/distrobuilder#830

Signed-off-by: JUN JIE NAN <nanjunjie@gmail.com>