images/gentoo.yaml: prepare for dracut-106's hostonly change
Quoting from the Gentoo News item:
"Starting with version 106 of sys-kernel/dracut the default for the "hostonly"
setting has changed from disabled to enabled when Dracut is invoked via
installkernel.

"hostonly" is a setting for Dracut that controls how much is included in the
generated initramfs image. When it is disabled Dracut aims to generate an
initramfs image that is bootable on any hardware. On the other hand, when this
setting is enabled, Dracut aims to generate an initramfs image containing only
what is needed to boot the current system. The advantage is a significantly
smaller initramfs images, but this comes with the cost of losing portability."

dracut-ng/dracut-ng#1158
https://www.gentoo.org/support/news-items/2025-02-03-dracut-hostonly.html

This change is relevant only for the vm variant images.

Signed-off-by: Joonas Niilola <juippis@gentoo.org>